+++ Tests/Support/Snapshots/edgecases-change-access-control/Sources/Entities/Pet.swift
@@ -5,7 +5,7 @@ import Foundation
 import NaiveDate
 
 struct Pet: Codable {
-    var id: Int?
+    var id: Int64?
     var category: Category?
     /// Example: "doggie"
     var name: String
@@ -21,7 +21,7 @@ struct Pet: Codable {
         case sold
     }
 
-    init(id: Int? = nil, category: Category? = nil, name: String, photoURLs: [String], tags: [Tag]? = nil, status: Status? = nil) {
+    init(id: Int64? = nil, category: Category? = nil, name: String, photoURLs: [String], tags: [Tag]? = nil, status: Status? = nil) {
         self.id = id
         self.category = category
         self.name = name
@@ -32,7 +32,7 @@ struct Pet: Codable {
 
     init(from decoder: Decoder) throws {
         let values = try decoder.container(keyedBy: StringCodingKey.self)
-        self.id = try values.decodeIfPresent(Int.self, forKey: "id")
+        self.id = try values.decodeIfPresent(Int64.self, forKey: "id")
         self.category = try values.decodeIfPresent(Category.self, forKey: "category")
         self.name = try values.decode(String.self, forKey: "name")
         self.photoURLs = try values.decode([String].self, forKey: "photoUrls")
