+++ Tests/Support/Snapshots/edgecases-coding-keys/Sources/Entities/FormatTest.swift
@@ -6,10 +6,10 @@ import NaiveDate
 
 public struct FormatTest: Codable {
     public var integer: Int?
-    public var int32: Int?
-    public var int64: Int?
+    public var int32: Int32?
+    public var int64: Int64?
     public var number: Double
-    public var float: Double?
+    public var float: Float?
     public var double: Double?
     public var string: String?
     public var byte: Data
@@ -19,7 +19,7 @@ public struct FormatTest: Codable {
     public var uuid: UUID?
     public var password: String
 
-    public init(integer: Int? = nil, int32: Int? = nil, int64: Int? = nil, number: Double, float: Double? = nil, double: Double? = nil, string: String? = nil, byte: Data, binary: String? = nil, date: NaiveDate, dateTime: Date? = nil, uuid: UUID? = nil, password: String) {
+    public init(integer: Int? = nil, int32: Int32? = nil, int64: Int64? = nil, number: Double, float: Float? = nil, double: Double? = nil, string: String? = nil, byte: Data, binary: String? = nil, date: NaiveDate, dateTime: Date? = nil, uuid: UUID? = nil, password: String) {
         self.integer = integer
         self.int32 = int32
         self.int64 = int64
