+++ Tests/Support/Snapshots/edgecases-indent-with-two-width-spaces/Sources/Entities/Tag.swift
@@ -5,17 +5,17 @@ import Foundation
 import NaiveDate
 
 public struct Tag: Codable {
-  public var id: Int?
+  public var id: Int64?
   public var name: String?
 
-  public init(id: Int? = nil, name: String? = nil) {
+  public init(id: Int64? = nil, name: String? = nil) {
     self.id = id
     self.name = name
   }
 
   public init(from decoder: Decoder) throws {
     let values = try decoder.container(keyedBy: StringCodingKey.self)
-    self.id = try values.decodeIfPresent(Int.self, forKey: "id")
+    self.id = try values.decodeIfPresent(Int64.self, forKey: "id")
     self.name = try values.decodeIfPresent(String.self, forKey: "name")
   }
